+++ src/main/java/com/jnape/palatable/lambda/adt/choice/Choice2.java
@@ -30,6 +30,11 @@ public abstract class Choice2<A, B> implements CoProduct2<A, B>, Applicative<B, 
     }
 
     @Override
+    public Choice2<B, A> invert() {
+        return match(Choice2::b, Choice2::a);
+    }
+
+    @Override
     @SuppressWarnings("unchecked")
     public final <C> Choice2<A, C> fmap(Function<? super B, ? extends C> fn) {
         return (Choice2<A, C>) Applicative.super.fmap(fn);